tobbacodeseasedetection/requirements.txt
# Web Framework
flask==3.0.0
gunicorn==21.2.0

# HTTP Requests
requests==2.31.0

# Firebase
firebase-admin==6.4.0

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# YOLO Model
ultralytics==8.0.220

# Environment Variables
python-dotenv==1.0.0

# Email Support
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
pytest-mock==3.12.0

# Rate Limiting (optional - for production with Redis)
# redis==5.0.1